Use Parallel and ConcurrentBag; Force Stars filter; Merge with contributors; Misc fixes & improvements
