+++ TeraFinder.Core/Resources/translations/lang_es.txt
@@ -4,9 +4,9 @@ Plugin.TeraViewer=Visor/Editor de Teraincursiones
 Plugin.SeedChecker=Inspector de Seeds
 Plugin.FlagEditor=Edita Flags del Juego
 Plugin.NewsImporter=Importar Noticias Poképortal
-Plugin.ImportNews=Import from files...
-Plugin.OutbreakNull=Import Empty (Null) Outbreak Event
-Plugin.RaidNull=Import Empty (Null) Raid Event
+Plugin.ImportNews=Importar desde archivos...
+Plugin.OutbreakNull=Importar Apariciones Masivas Vacías (Evento Nulo)
+Plugin.RaidNull=Importar Teraincursión Vacía (Evento Nulo)
 Plugin.OutbreakViewer=Visor/Editor de Apariciones Masivas
 Plugin.MapPaldea=Paldea
 Plugin.MapKitakami=Noroteo
@@ -19,8 +19,8 @@ TeraFinderForm=Tera Finder
 TeraFinderForm.remoteConnectToolStripMenuItem=Herramientas
 TeraFinderForm.btnRemoteConnect=Conectar a Dispositivo Remoto
 TeraFinderForm.languageToolStrip=Idioma Por Defecto
-TeraFinderForm.importNullRaid=Import Empty (Null) Raid Event
-TeraFinderForm.importNullOutbreak=Import Empty (Null) Outbreak Event
+TeraFinderForm.importNullRaid=Importar Teraincursión Vacía (Evento Nulo)
+TeraFinderForm.importNullOutbreak=Importar Apariciones Masivas Vacías (Evento Nulo)
 TeraFinderForm.grpSAV=Archivo SAV
 TeraFinderForm.grpSavTools=Herramientas de SAV
 TeraFinderForm.grpStaticTools=Herramientas Independientes
@@ -61,7 +61,7 @@ OutbreakForm.DeviceDisconnected=Dispositivo desconectado.
 OutbreakForm.ErrorParsing=Error al analizar:
 OutbreakForm.LoadDefault=¿Cargar datos legales por defecto para {species}?
 OutbreakForm.SpeciesExclusive=¡{species} es exclusivo de {game}!
-OutbreakForm.lblMap=Location Map:
+OutbreakForm.lblMap=Mapa:
 
 CalculatorForm=Calculadora de Teraincursiones
 CalculatorForm.lblSeed=Seed:
@@ -85,7 +85,7 @@ CalculatorForm.lblSpd=SpD:
 CalculatorForm.lblSpe=Spe:
 CalculatorForm.lblStars=Estrella(s):
 CalculatorForm.lblSpecies=Especie:
-CalculatorForm.lblTera=Tipo Tera:
+CalculatorForm.lblTera=Teratipo:
 CalculatorForm.lblEC=EC:
 CalculatorForm.lblScale=Escala:
 CalculatorForm.lblAbility=Habilidad:
@@ -103,8 +103,8 @@ CalculatorForm.btnSavePk9=Guardar Resultado Seleccionado como PK9
 CalculatorForm.btnToPkmEditor=Enviar Resultado Seleccionado a PKM Editor
 CalculatorForm.btnSendToEditor=Enviar Resultado Seleccionado a Raid Editor
 CalculatorForm.lblEventct=Índice Evento:
-CalculatorForm.lblMap=Location:
-CalculatorForm.btnCopySeed=Copy Seed
+CalculatorForm.lblMap=Ubicación:
+CalculatorForm.btnCopySeed=Copiar Seed
 
 RewardCalcForm=Calculadora de Recompensas
 RewardCalcForm.grpFilters=Filtros
@@ -125,7 +125,7 @@ RewardCalcForm.grpSettings=Ajustes de Teraincursión
 RewardCalcForm.lblSeed=Seed:
 RewardCalcForm.lblCalcs=Max Calc:
 RewardCalcForm.lblContent=Contenido:
-RewardCalcForm.lblSandwich=Poder Sandwich:
+RewardCalcForm.lblSandwich=Poder Sándwich:
 RewardCalcForm.grpSearch=Búsqueda
 RewardCalcForm.lblFound=Encontrado:
 RewardCalcForm.chkAllResults=Mostrar todos los resultados (?)
@@ -134,8 +134,8 @@ RewardCalcForm.btnSaveAllTxt=Guardar resultados como TXT
 RewardCalcForm.btnSaveSelectedTxt=Guardar resultados seleccionados como TXT
 RewardCalcForm.btnSendSelectedRaid=Enviar Resultado Seleccionado a Raid Editor
 RewardCalcForm.lblEventCt=Índice Evento:
-RewardCalcForm.lblMap=Location:
-RewardCalcForm.btnCopySeed=Copy Seed
+RewardCalcForm.lblMap=Ubicación:
+RewardCalcForm.btnCopySeed=Copiar Seed
 
 CheckerForm=Inspector de Seeds
 CheckerForm.lblTID=TID:
@@ -162,19 +162,19 @@ ConnectionForm.radioWiFi=Wi-Fi
 ConnectionForm.radioUSB=USB
 ConnectionForm.lblAddress=Dirección:
 ConnectionForm.lblPort=Puerto:
-ConnectionForm.chkOutbreaksMain=Descargar Paldea Apariciones Masivas (?)
-ConnectionForm.chkOutbreaksDLC=Descargar Noroteo Apariciones Masivas (?)
-ConnectionForm.chkOutbreaksDLC2=Descargar Arándano Apariciones Masivas (?)
+ConnectionForm.chkOutbreaksMain=Descargar Apariciones Masivas (Paldea) (?)
+ConnectionForm.chkOutbreaksDLC=Descargar Apariciones Masivas (Noroteo) (?)
+ConnectionForm.chkOutbreaksDLC2=Descargar Apariciones Masivas (Arándano) (?)
 ConnectionForm.btnConnect=Conectar
 ConnectionForm.btnDisconnect=Desconectar
 
 EditorForm=Editor de Teraincursiones
-EditorForm.toolsToolStripMenuItem=Tools
+EditorForm.toolsToolStripMenuItem=Herramientas
 EditorForm.btnShinifyAll=Todas las raids Variocolor (Al azar)
 EditorForm.btnShinifyAllEncounters=Todas las raids Variocolor (Mantener)
-EditorForm.btnShinifyCurrent=Shinify Current Raid
-EditorForm.btnRandomizeCurrent=Randomize Current Raid
-EditorForm.btnRandomizeAll=Randomize All Raids
+EditorForm.btnShinifyCurrent=Volver Variocolor Raid Actual
+EditorForm.btnRandomizeCurrent=Aleatorizar Raid Actual
+EditorForm.btnRandomizeAll=Aleatorizar todas las Raids
 EditorForm.GrpRaidInfo=Info. de Teraincursiones
 EditorForm.lblSeed=Seed:
 EditorForm.lblContent=Contenido:
@@ -189,7 +189,7 @@ EditorForm.lblSpD=SpD:
 EditorForm.lblSpe=Spe:
 EditorForm.lblSpecies=Especie:
 EditorForm.lblAbility=Habilidad:
-EditorForm.lblTera=Tipo Tera:
+EditorForm.lblTera=Teratipo:
 EditorForm.lblNature=Naturaleza:
 EditorForm.lblShiny=Variocolor:
 EditorForm.lblGender=Género:
@@ -205,8 +205,8 @@ EditorForm.btnOpenCalculator=Calculadora Teraincursiones
 EditorForm.btnOpenRewardCalculator=Calculadora Recompensas
 EditorForm.LblLvl=Nvl.
 EditorForm.CmbRaid=Teraincursión
-EditorForm.lblMap=Location Map:
-EditorForm.toolTipMoves=Extra Raid Moves:
+EditorForm.lblMap=Mapa:
+EditorForm.toolTipMoves=Movimientos Extra:
 
 ShinifyForm.lblValue=Progreso:
 
@@ -219,10 +219,10 @@ ProgressForm.chkDefeated=Derrotado
 ProgressForm.btnApplyRaid7=Aplicar
 
 RewardListForm=Recompensas
-RewardListForm.grpNoBoost=Sin Poder Sandwich
-RewardListForm.grp1Boost=Poder Sandwich Nv. 1
-RewardListForm.grp2Boost=Poder Sandwich Nv. 2
-RewardListForm.grp3Boost=Poder Sandwich Nv. 3
+RewardListForm.grpNoBoost=Sin Poder Sándwich
+RewardListForm.grp1Boost=Poder Sándwich Nv. 1
+RewardListForm.grp2Boost=Poder Sándwich Nv. 2
+RewardListForm.grp3Boost=Poder Sándwich Nv. 3
 
 ImportNews.Title=Abrir archivo .ZIP o carpeta con Eventos de Poképortal
 ImportNews.Filter=Todos los archivos
@@ -235,7 +235,7 @@ GameVersionSV=ScVi
 GameVersionSL=Escarlata
 GameVersionVL=Púrpura
 TrainerBlank=TeraFinder
-NewsEvent=Evento Poképortal: [Raid: {0}, Outbreak: {1}]
+NewsEvent=Evento Poképortal: [Raid: {0}, A. Masivas: {1}]
 None=Ninguno
 SAVInvalid=Archivo SAV no es válido.
 MsgSuccess=Listo.
@@ -250,7 +250,7 @@ FiltersPopup=¿Quieres aplicar los filtros en la búsqueda actual?
 FiltersApply=Aplicar Filtros
 ActionSearch=Buscar
 ActionStop=Parar
-ToolTipAccurate=Forzar la calculadora a determinar el Tipo Tera y variocolor de una seed dada, para así determinar con precisión el tipo de Teralitos e información extra. Esto hace las búsquedas un poco lentas.
+ToolTipAccurate=Forzar la calculadora a definir el Teratipo y variocolor de una seed dada, para así determinar con precisión el tipo de Teralitos e información extra. Esto hace las búsquedas un poco lentas.
 ToolTipAllResults=Deshabilitado - Detener cada hilo de búsqueda al primer resultado que coincida con los filtros.  Habilitado - Computa todos los resultados posibles hasta que el número en Max Calc sea alcanzado.  Se ignora si no hay ningún filtro establecido.
 
 ActionConnect=Conectar
@@ -276,11 +276,11 @@ TeraShiny.CmbStar=Sólo Estrella
 TeraShiny.CmbSquare=Sólo Cuadrado
 
 GameProgress.Beginning=Comienzo
-GameProgress.UnlockedTeraRaids=Teraincursiones Desbloqueadas
-GameProgress.Unlocked3Stars=Desbloqueadas 3 Estrellas
-GameProgress.Unlocked4Stars=Desbloqueadas 4 Estrellas
-GameProgress.Unlocked5Stars=Desbloqueadas 5 Estrellas
-GameProgress.Unlocked6Stars=Desbloqueadas 6 Estrellas
+GameProgress.UnlockedTeraRaids=Teraincursiones Habilitadas
+GameProgress.Unlocked3Stars=3 Estrellas Disponibles
+GameProgress.Unlocked4Stars=4 Estrellas Disponibles
+GameProgress.Unlocked5Stars=5 Estrellas Disponibles
+GameProgress.Unlocked6Stars=6 Estrellas Disponibles
 GameProgress.None=Ninguno
 
 RaidContent.Standard=Estándar
